Modify: 삭제 버튼 짤 카드에 추가
